add 5 round test and output statistics

+++ src/com/scu/coen383/team2/scheduling/ScheduleBase.java
@@ -9,7 +9,7 @@ public abstract class ScheduleBase {
     private Stats stats = new Stats();
     public class Stats
     {
-        private int turnaroundTime;
+        private float turnaroundTime;
         private int waitingTime;
         private int responseTime;
         private int processCount;
